same as last commit

+++ ft_putptr.c
@@ -6,7 +6,7 @@
 /*   By: yemoreno <yemoreno@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 17:52:23 by yemoreno          #+#    #+#             */
-/*   Updated: 2023/10/12 20:02:57 by yemoreno         ###   ########.fr       */
+/*   Updated: 2023/10/12 20:10:48 by yemoreno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	ft_putptr(void *ptr, int *count)
 {
-	unsigned long	ptr_address;
+	unsigned int	ptr_address;
 
-	ptr_address = (unsigned long) ptr;
+	ptr_address = (unsigned int) ptr;
 	ft_putstr("0x", count);
 	ft_putnbr_base_unsigned(ptr_address, HEX_LOW_BASE, count);
 }
